requirements.txt
Django==3.2.13
asgiref==3.5.2
pip==22.1.2
pytz==2022.1
setuptools==57.4.0
sqlparse==0.4.2
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pytest==7.1.2